Extract platform in sbt-bloop
The extraction is done simply by checking whether the Scala Native or
Scala.js is enabled. If none of them is enabled, then the platform is
JVM.

It could be interesting to use sbt-cross-project to check the platform,
but it is possible to create a Scala Native or Scala.js project without
using sbt-cross-project. The current solution has the advantage of
working in either cases.

In the maven plugin, the platform is always chosen to be JVM.