Add test to ensure that recompilation in nailgun is no-op
This test is important to prove that killing the server and restarting
it again should still pick up the analysis files correctly and don't
recompile anything if there haven't been any changes.

This commit makes also sure that we can run these tests in bloop. There
were some problems when doing so regarding in, out and error streams
which are fixed in `NailgunTestUtils`.